Update and refactor ArnaButtons
