+++ lib/arna.dart
@@ -36,15 +36,6 @@ export 'package:arna/src/base/tab_item.dart';
 export 'package:arna/src/base/tab_view.dart';
 export 'package:arna/src/base/text_selection.dart';
 export 'package:arna/src/base/tooltip.dart';
-export 'package:arna/src/buttons/back_button.dart';
-export 'package:arna/src/buttons/borderless_button.dart';
-export 'package:arna/src/buttons/button.dart';
-export 'package:arna/src/buttons/close_button.dart';
-export 'package:arna/src/buttons/color_button.dart';
-export 'package:arna/src/buttons/icon_button.dart';
-export 'package:arna/src/buttons/linked_buttons.dart';
-export 'package:arna/src/buttons/pill_button.dart';
-export 'package:arna/src/buttons/text_button.dart';
 export 'package:arna/src/containers/badge.dart';
 export 'package:arna/src/containers/card.dart';
 export 'package:arna/src/containers/expansion_panel.dart';
@@ -55,6 +46,11 @@ export 'package:arna/src/containers/header_bar/sliver_header_bar.dart';
 export 'package:arna/src/containers/list.dart';
 export 'package:arna/src/containers/list_tile.dart';
 export 'package:arna/src/containers/reorderable_list.dart';
+export 'package:arna/src/controls/buttons/back_button.dart';
+export 'package:arna/src/controls/buttons/button.dart';
+export 'package:arna/src/controls/buttons/close_button.dart';
+export 'package:arna/src/controls/buttons/color_button.dart';
+export 'package:arna/src/controls/buttons/linked_buttons.dart';
 export 'package:arna/src/dialogs/about_dialog.dart';
 export 'package:arna/src/dialogs/alert_dialog.dart';
 export 'package:arna/src/dialogs/dialog.dart';
