+++ lib/src/base/base_widget.dart
@@ -27,8 +27,6 @@ typedef ArnaBaseWidgetBuilder = Widget Function(
 /// See also:
 ///
 ///  * [ArnaButton]
-///  * [ArnaIconButton]
-///  * [ArnaTextButton]
 ///  * [ArnaMasterItem]
 ///  * [ArnaSideBarItem]
 ///  * [ArnaLinkedButtons]
