+++ lib/src/base/master_detail_scaffold.dart
@@ -34,7 +34,7 @@ class ArnaMasterDetailScaffold extends StatefulWidget {
 
   /// A list of Widgets to display in a row after the [title] widget.
   ///
-  /// Typically these widgets are [ArnaIconButton]s representing common operations. For less common operations,
+  /// Typically these widgets are [ArnaButton.icon]s representing common operations. For less common operations,
   /// consider using an [ArnaPopupMenuButton] as the last action.
   ///
   /// The [actions] become the trailing component of the [NavigationToolbar] built by this widget. The height of each
@@ -156,7 +156,7 @@ class _LateralPage extends StatefulWidget {
 
   /// A list of Widgets to display in a row after the [title] widget.
   ///
-  /// Typically these widgets are [ArnaIconButton]s representing common operations. For less common operations,
+  /// Typically these widgets are [ArnaButton.icon]s representing common operations. For less common operations,
   /// consider using an [ArnaPopupMenuButton] as the last action.
   ///
   /// The [actions] become the trailing component of the [NavigationToolbar] built by this widget. The height of each
@@ -319,7 +319,7 @@ class _NestedPage extends StatefulWidget {
 
   /// A list of Widgets to display in a row after the [title] widget.
   ///
-  /// Typically these widgets are [ArnaIconButton]s representing common operations. For less common operations,
+  /// Typically these widgets are [ArnaButton.icon]s representing common operations. For less common operations,
   /// consider using an [ArnaPopupMenuButton] as the last action.
   ///
   /// The [actions] become the trailing component of the [NavigationToolbar] built by this widget. The height of each
@@ -574,7 +574,7 @@ class MasterNavigationItem {
 
   /// A list of Widgets to display in a row after the [title] widget.
   ///
-  /// Typically these widgets are [ArnaIconButton]s representing common operations. For less common operations,
+  /// Typically these widgets are [ArnaButton.icon]s representing common operations. For less common operations,
   /// consider using an [ArnaPopupMenuButton] as the last action.
   ///
   /// The [actions] become the trailing component of the [NavigationToolbar] built by this widget. The height of each
