+++ lib/src/base/scaffold.dart
@@ -29,7 +29,7 @@ class ArnaScaffold extends StatelessWidget {
 
   /// A list of Widgets to display in a row after the [title] widget.
   ///
-  /// Typically these widgets are [ArnaIconButton]s representing common operations. For less common operations,
+  /// Typically these widgets are [ArnaButton.icon]s representing common operations. For less common operations,
   /// consider using an [ArnaPopupMenuButton] as the last action.
   ///
   /// The [actions] become the trailing component of the [NavigationToolbar] built by this widget. The height of each
