--- lib/src/buttons/text_button.dart
@@ -1,85 +1,0 @@
-import 'package:arna/arna.dart';
-
-/// An Arna-styled text button.
-class ArnaTextButton extends StatelessWidget {
-  /// Creates a text button.
-  const ArnaTextButton({
-    super.key,
-    required this.label,
-    required this.onPressed,
-    this.onLongPress,
-    this.tooltipMessage,
-    this.buttonType = ButtonType.normal,
-    this.buttonSize = ButtonSize.normal,
-    this.isFocusable = true,
-    this.autofocus = false,
-    this.accentColor,
-    this.cursor = MouseCursor.defer,
-    this.semanticLabel,
-    this.enableFeedback = true,
-  });
-
-  /// The text label of the button.
-  final String? label;
-
-  /// The callback that is called when a button is tapped.
-  ///
-  /// If this callback and [onLongPress] are null, then the button will be disabled.
-  final VoidCallback? onPressed;
-
-  /// The callback that is called when a button is long-pressed.
-  ///
-  /// If this callback and [onPressed] are null, then the button will be disabled.
-  final VoidCallback? onLongPress;
-
-  /// Text that describes the action that will occur when the button is pressed.
-  final String? tooltipMessage;
-
-  /// The type of the button.
-  final ButtonType buttonType;
-
-  /// The size of the button.
-  final ButtonSize buttonSize;
-
-  /// Whether this button is focusable or not.
-  final bool isFocusable;
-
-  /// Whether this button should focus itself if nothing else is already focused.
-  final bool autofocus;
-
-  /// The color of the button's focused border.
-  final Color? accentColor;
-
-  /// The cursor for a mouse pointer when it enters or is hovering over the button.
-  final MouseCursor cursor;
-
-  /// The semantic label of the button.
-  final String? semanticLabel;
-
-  /// Whether detected gestures should provide acoustic and/or haptic feedback.
-  ///
-  /// For example, on Android a long-press will produce a short vibration, when feedback is enabled.
-  ///
-  /// See also:
-  ///
-  ///  * [ArnaFeedback] for providing platform-specific feedback to certain actions.
-  final bool enableFeedback;
-
-  @override
-  Widget build(BuildContext context) {
-    return ArnaButton(
-      label: label,
-      onPressed: onPressed,
-      onLongPress: onLongPress,
-      tooltipMessage: tooltipMessage,
-      buttonType: buttonType,
-      buttonSize: buttonSize,
-      isFocusable: isFocusable,
-      autofocus: autofocus,
-      accentColor: accentColor,
-      cursor: cursor,
-      semanticLabel: semanticLabel,
-      enableFeedback: enableFeedback,
-    );
-  }
-}
