+++ lib/src/containers/header_bar/header_bar.dart
@@ -37,7 +37,7 @@ class ArnaHeaderBar extends StatelessWidget {
 
   /// A list of Widgets to display in a row after the [middle] widget.
   ///
-  /// Typically these widgets are [ArnaIconButton]s representing common
+  /// Typically these widgets are [ArnaButton.icon]s representing common
   /// operations. For less common operations, consider using an
   /// [ArnaPopupMenuButton] as the last action.
   ///
