+++ lib/src/controls/buttons/close_button.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart' show MaterialLocalizations;
 
 /// An Arna-styled close button.
 ///
-/// An [ArnaCloseButton] is an [ArnaBorderlessButton] with a "close" icon. When pressed, the close button calls
-/// [Navigator.maybePop] to return to the previous route.
+/// An [ArnaCloseButton] is an [ArnaButton] with a "close" icon. When pressed,
+/// the close button calls [Navigator.maybePop] to return to the previous route.
 ///
 /// Use an [ArnaCloseButton] instead of an [ArnaBackButton] on dialogs.
 class ArnaCloseButton extends StatelessWidget {
@@ -14,25 +14,23 @@ class ArnaCloseButton extends StatelessWidget {
     this.onPressed,
   });
 
-  /// An override callback to perform instead of the default behavior which is to pop the [Navigator].
+  /// An override callback to perform instead of the default behavior which is
+  /// to pop the [Navigator].
   ///
-  /// It can, for instance, be used to pop the platform's navigation stack via [SystemNavigator] instead of Flutter's
-  /// [Navigator] in add-to-app situations.
+  /// It can, for instance, be used to pop the platform's navigation stack via
+  /// [SystemNavigator] instead of Flutter's [Navigator] in add-to-app
+  /// situations.
   ///
   /// Defaults to null.
   final VoidCallback? onPressed;
 
   @override
   Widget build(BuildContext context) {
-    return ArnaBorderlessButton(
+    return ArnaButton(
       icon: Icons.close_outlined,
-      onPressed: () {
-        if (onPressed != null) {
-          onPressed!();
-        } else {
-          Navigator.maybePop(context);
-        }
-      },
+      onPressed: () =>
+          onPressed != null ? onPressed!() : Navigator.maybePop(context),
+      buttonType: ButtonType.borderless,
       tooltipMessage: MaterialLocalizations.of(context).closeButtonTooltip,
       semanticLabel: MaterialLocalizations.of(context).closeButtonTooltip,
     );
