+++ lib/src/dialogs/popup_dialog.dart
@@ -10,11 +10,11 @@ import 'package:flutter/material.dart' show MaterialLocalizations;
 ///
 /// {@tool snippet}
 ///
-/// This snippet shows an ArnaIconButton which, when pressed, displays a dialog box in desktop and tablet mode and a
+/// This snippet shows an ArnaButton.icon which, when pressed, displays a dialog box in desktop and tablet mode and a
 /// new page on phones.
 ///
 /// ```dart
-/// ArnaIconButton(
+/// ArnaButton.icon(
 ///     icon: Icons.info_outlined,
 ///     onPressed: () => showArnaPopupDialog(
 ///         context: context,
@@ -54,7 +54,7 @@ class _ArnaPopupDialog extends StatelessWidget {
 
   /// A list of Widgets to display in a row after the [title] widget.
   ///
-  /// Typically these widgets are [ArnaIconButton]s representing common operations. For less common operations,
+  /// Typically these widgets are [ArnaButton.icon]s representing common operations. For less common operations,
   /// consider using an [ArnaPopupMenuButton] as the last action.
   ///
   /// The [actions] become the trailing component of the [NavigationToolbar] built by this widget. The height of each
@@ -114,11 +114,11 @@ class _ArnaPopupDialog extends StatelessWidget {
 ///
 /// {@tool snippet}
 ///
-/// This snippet shows an ArnaIconButton which, when pressed, displays a dialog box in desktop and tablet mode and a
+/// This snippet shows an ArnaButton.icon which, when pressed, displays a dialog box in desktop and tablet mode and a
 /// new page on phones.
 ///
 /// ```dart
-/// ArnaIconButton(
+/// ArnaButton.icon(
 ///     icon: Icons.info_outlined,
 ///     onPressed: () => showArnaPopupDialog(
 ///         context: context,
@@ -159,7 +159,7 @@ class _ArnaPopupPage extends StatelessWidget {
 
   /// A list of Widgets to display in a row after the [title] widget.
   ///
-  /// Typically these widgets are [ArnaIconButton]s representing common operations. For less common operations,
+  /// Typically these widgets are [ArnaButton.icon]s representing common operations. For less common operations,
   /// consider using an [ArnaPopupMenuButton] as the last action.
   ///
   /// The [actions] become the trailing component of the [NavigationToolbar] built by this widget. The height of each
